# contracts/tester/v1/openapi.yaml
openapi: 3.0.3
info:
  title: TesterService API
  version: 0.0.1
paths:
...
  /contests:
    get:
      operationId: ListContests
      parameters:
        - name: page
          in: query
          required: true
          schema:
            type: integer
            format: int32
          example: 1
        - name: pageSize
          in: query
          required: true
          schema:
            type: integer
            format: int32
          example: 10
      security:
        - bearerAuth: [ ]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListContestsResponse'
    post:
      operationId: CreateContest
      security:
        - bearerAuth: [ ]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateContestResponse'
  /contests/{id}:
    get:
      operationId: GetContest
      security:
        - bearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
          example: 1
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetContestResponse'
    delete:
      operationId: DeleteContest
      security:
        - bearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
          example: 3
      responses:
        "200":
          description: OK
    patch:
      operationId: UpdateContest
      security:
        - bearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
          example: 3
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateContestRequest'
      responses:
        "200":
          description: OK
...
  /participants:
    get:
      operationId: ListParticipants
      security:
        - bearerAuth: [ ]
      parameters:
        - name: contest_id
          in: query
          required: true
          schema:
            type: integer
            format: int32
          example: 1
        - name: page
          in: query
          required: true
          schema:
            type: integer
            format: int32
          example: 1
        - name: pageSize
          in: query
          required: true
          schema:
            type: integer
            format: int32
          example: 10
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListParticipantsResponse'
    post:
      operationId: CreateParticipant
      security:
        - bearerAuth: [ ]
      parameters:
        - name: contest_id
          in: query
          required: true
          schema:
            type: integer
            format: int32
          example: 1
        - name: user_id
          in: query
          required: true
          schema:
            type: integer
            format: int32
          example: 1
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateParticipantResponse'
    delete:
      operationId: DeleteParticipant
      security:
        - bearerAuth: [ ]
      parameters:
        - name: participant_id
          in: query
          required: true
          schema:
            type: integer
            format: int32
          example: 1
      responses:
        "200":
          description: OK
    patch:
      operationId: UpdateParticipant
      security:
        - bearerAuth: [ ]
      parameters:
        - name: participant_id
          in: query
          required: true
          schema:
            type: integer
            format: int32
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateParticipantRequest'
      responses:
        "200":
          description: OK
...
  /auth/terminate:
    post:
      operationId: Terminate
      security:
        - bearerAuth: [ ]
      responses:
        "200":
          description: OK
          content: { }
...
  /auth/refresh:
    post:
      operationId: Refresh
      security:
        - bearerAuth: [ ]
      responses:
        "200":
          description: OK
          content: { }
  /auth/sessions:
    get:
      operationId: ListSessions
      security:
        - bearerAuth: [ ]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListSessionsResponse'
...
components:
  schemas:
    Pagination:
      type: object
      required:
        - page
        - total
      properties:
        page:
          type: integer
          format: int32
          example: 1
        total:
          type: integer
          format: int32
          example: 10
...
    Contest:
      type: object
      required:
        - id
        - title
        - created_at
        - updated_at
      properties:
        id:
          type: integer
          format: int32
          example: 1
        title:
          type: string
          example: "Test contest"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ContestsListItem:
      type: object
      required:
        - id
        - title
        - created_at
        - updated_at
      properties:
        id:
          type: integer
          format: int32
          example: 1
        title:
          type: string
          example: "Test contest"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ListContestsResponse:
      type: object
      required:
        - contests
        - pagination
      properties:
        contests:
          type: array
          items:
            $ref: '#/components/schemas/ContestsListItem'
        pagination:
          $ref: '#/components/schemas/Pagination'
    CreateContestResponse:
      type: object
      required:
        - id
      properties:
        id:
          type: integer
          format: int32
          example: 1
    TasksListItem:
      type: object
      required:
        - id
        - problem_id
        - position
        - title
        - memory_limit
        - time_limit
        - created_at
        - updated_at
      properties:
        id:
          type: integer
          format: int32
          example: 1
        problem_id:
          type: integer
          format: int32
          example: 1
        position:
          type: integer
          format: int32
          example: 1
        title:
          type: string
          example: "Test task"
        memory_limit:
          type: integer
          format: int32
          example: 256
        time_limit:
          type: integer
          format: int32
          example: 1000
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    GetContestResponse:
      type: object
      required:
        - contest
        - tasks
      properties:
        contest:
          $ref: '#/components/schemas/Contest'
        tasks:
          type: array
          items:
            type: object
            required:
              - task
              - solution
            properties:
              task:
                $ref: '#/components/schemas/TasksListItem'
              solution:
                $ref: '#/components/schemas/SolutionsListItem'
    CreateParticipantResponse:
      type: object
      required:
        - id
      properties:
        id:
          type: integer
          format: int32
          example: 1
...
    ParticipantsListItem:
      type: object
      required:
        - id
        - name
        - user_id
        - created_at
        - updated_at
      properties:
        id:
          type: integer
          format: int32
          example: 1
        name:
          type: string
          example: "Test user"
        user_id:
          type: integer
          format: int32
          example: 1
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ListParticipantsResponse:
      type: object
      required:
        - participants
        - pagination
      properties:
        participants:
          type: array
          items:
            $ref: '#/components/schemas/ParticipantsListItem'
        pagination:
          $ref: '#/components/schemas/Pagination'
...
    UpdateContestRequest:
      type: object
      properties:
        title:
          type: string
          example: "Test contest"
    UpdateParticipantRequest:
      type: object
      properties:
        name:
          type: string
          example: "Test user"
...
    SolutionsListItem:
      type: object
      required:
        - id

        - participant_id
        - participant_name

        - state
        - score
        - penalty
        - time_stat
        - memory_stat
        - language

        - task_id
        - task_position
        - task_title

        - contest_id
        - contest_title

        - updated_at
        - created_at
      properties:
        id:
          type: integer
          format: int32
          example: 1

        participant_id:
          type: integer
          format: int32
          example: 1
        participant_name:
          type: string
          example: "user123"

        state:
          type: integer
          format: int32
          example: 1
        score:
          type: integer
          format: int32
          example: 100
        penalty:
          type: integer
          format: int32
          example: 28
        time_stat:
          type: integer
          format: int32
          example: 1000
        memory_stat:
          type: integer
          format: int32
          example: 256
        language:
          type: integer
          format: int32
          example: 1

        task_id:
          type: integer
          format: int32
          example: 1
        task_position:
          type: integer
          format: int32
          example: 1
        task_title:
          type: string
          example: "Task title"

        contest_id:
          type: integer
          format: int32
          example: 1
        contest_title:
          type: string
          example: "Contest title"

        updated_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
...
    Session:
      type: object
      required:
        - id
        - userId
        - role
        - createdAt
        - expiresAt
        - userAgent
        - ip
      properties:
        id:
          type: string
        userId:
          type: integer
          format: int32
          example: 13
        role:
          type: integer
          format: int32
          example: 1
        createdAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
        userAgent:
          type: string
          example: Mozilla/5.0
        ip:
          type: string
          example: 154.23.50.1
    ListSessionsResponse:
      type: object
      required:
        - sessions
      properties:
        sessions:
          type: array
          items:
            $ref: '#/components/schemas/Session'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT